@@ -1,14 +1,13 @@
 package com.anew.note.adapter;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 import com.anew.note.R;
 import com.anew.note.model.TipModel;
-import com.anew.note.utils.SPUtils;
 
 import java.util.ArrayList;
 
